Fixed Shooterspin in RobotBuilder so stops generating isFinished returning a value
Changed the definition of ShooterSpin in Robot Builder from a SetPoint
Command to just a Command (actually deleted the old one and created a
new one by the same name).  Now it does not  generate a automatic return
value on isFinished() so we can always have it seeking the PID set
speed.